breaking change: remove *Owned Deref
The Owned PR was broken. The lifetime's of the Cow fields are wrong.
The Cow str fields of an owned proto can end up referring to garbage if
`field.to_owned()` is called and then the original field is modified or
the Owned proto is dropped.

Example:

```rust
let proto = FooOwned::try_from(&buf)?;
let owned_copy = proto.str_field.to_owned();
drop(proto);
println!("this code should not be valid {}", owned_copy);
```

`owned_copy` refers to nothing now.

The fix _seems_ to be to tie the transmuted lifetime of the proto from
`'static` to the lifetime of `&self`.